+++ CHANGES.md
@@ -5,6 +5,7 @@
 ### Transformers
 - New `ReplaceReturns` transformer that replaces return statements (such as `[Return]` setting or `Return From Keyword` keyword) with `RETURN` statement ([#231](https://github.com/MarketSquare/robotframework-tidy/issues/231))
 - New `InlineIf` transformer that replaces IF blocks with inline IF ([#230](https://github.com/MarketSquare/robotframework-tidy/issues/230))
+- New `ReplaceBreakContinue` transformer that replaces `Continue For Loop` and `Exit For Loop` keyword variants with CONTINUE and BREAK statements ([#241](https://github.com/MarketSquare/robotframework-tidy/issues/241))
 
 ## 1.6.2